refactor: keep the path canonicalizer in the one branch that needs it
The container resolves the path for every request it routes, so the service is handed a path
that is already decoded and normalized and can match it as it stands. It goes back to being a
pure matcher, and its contract now says so.

The filter keeps a canonicalizer for its fallback, which is the one path here that nothing
upstream has resolved. It decodes before it removes dot segments, which is the order the
container applies.
