+++ client-cache-control-bundle/src/test/java/org/jahia/bundles/cache/client/filter/ClientCacheFilterTest.java
@@ -23,6 +23,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertSame;
 
 /**
  * The path the filter hands to the service for rule matching.
@@ -33,10 +34,12 @@ public class ClientCacheFilterTest {
 
     @Test
     public void theResolvedPathIsTheOneTheContainerChoseTheServletWith() {
-        // A prefix mapping splits the path in two: the mapped prefix, then the remainder. Together they
-        // are the path the container resolved, already decoded and normalized.
-        assertEquals("/modules/healthcheck", ClientCacheFilter.resolvedPath(request("/modules", "/healthcheck", "/modules/healthchec%6b")));
-        assertEquals("/modules/tools/index.jsp", ClientCacheFilter.resolvedPath(request("/modules", "/tools/index.jsp", "/modules//tools/index.jsp")));
+        // A prefix mapping splits the path in two: the mapped prefix, then the remainder. The container
+        // decoded and normalized both, so they are taken as they stand however the request was written.
+        assertEquals("/modules/healthcheck",
+                ClientCacheFilter.resolvedPath(request("/modules", "/healthcheck", "/modules/healthchec%6b")));
+        assertEquals("/modules/tools/index.jsp",
+                ClientCacheFilter.resolvedPath(request("/modules", "/tools/index.jsp", "/modules//tools/index.jsp")));
     }
 
     @Test
@@ -46,8 +49,46 @@ public class ClientCacheFilterTest {
 
     @Test
     public void theRequestUriAnswersWhenTheContainerExposesNeither() {
-        assertEquals("/sites/mysite/home.html", ClientCacheFilter.resolvedPath(request("", null, "/sites/mysite/home.html")));
-        assertEquals("/sites/mysite/home.html", ClientCacheFilter.resolvedPath(request(null, null, "/sites/mysite/home.html")));
+        // Nothing upstream has decoded or normalized this one, so the filter does it here. Each of these
+        // reaches the same resource, so each yields the one path the rules are matched on.
+        assertEquals("/modules/healthcheck", ClientCacheFilter.resolvedPath(request("", null, "/modules/healthchec%6b")));
+        assertEquals("/modules/healthcheck", ClientCacheFilter.resolvedPath(request(null, null, "/modules/%68ealthcheck")));
+        assertEquals("/modules/healthcheck", ClientCacheFilter.resolvedPath(request("", null, "/modules//healthcheck")));
+        assertEquals("/modules/tools/index.jsp", ClientCacheFilter.resolvedPath(request("", null, "/modules//tools/index.jsp")));
+        assertEquals("/modules/healthcheck", ClientCacheFilter.resolvedPath(request("", null, "/modules/./healthcheck")));
+        assertEquals("/modules/healthcheck", ClientCacheFilter.resolvedPath(request("", null, "/modules/x/../healthcheck")));
+    }
+
+    @Test
+    public void canonicalizeRewritesEverySpellingToOneForm() {
+        assertEquals("/modules/healthcheck", ClientCacheFilter.canonicalize("/modules/healthchec%6b"));
+        assertEquals("/modules/healthcheck", ClientCacheFilter.canonicalize("/modules//healthcheck"));
+        assertEquals("/modules/healthcheck", ClientCacheFilter.canonicalize("/modules/./healthcheck"));
+        assertEquals("/modules/healthcheck", ClientCacheFilter.canonicalize("/modules/x/../healthcheck"));
+        assertEquals("/modules/tools/index.jsp", ClientCacheFilter.canonicalize("/modules/tool%73//index.jsp"));
+        // Decoding runs before dot segments are removed, which is the order the container applies, so an
+        // encoded dot segment resolves the same way a written one does.
+        assertEquals("/healthcheck", ClientCacheFilter.canonicalize("/modules/%2e%2e/healthcheck"));
+        // A multi-byte character written as several percent groups decodes to the character it stands for.
+        assertEquals("/files/été.pdf", ClientCacheFilter.canonicalize("/files/%C3%A9t%C3%A9.pdf"));
+    }
+
+    @Test
+    public void canonicalizeLeavesAPathThatIsAlreadyCanonicalUntouched() {
+        String uri = "/modules/ckeditor/javascript/ckeditor.js";
+        assertSame(uri, ClientCacheFilter.canonicalize(uri));
+        // A percent sign that is not followed by two hexadecimal digits belongs to the name and is kept as
+        // it stands. An encoded one is decoded, like any other encoded byte.
+        assertEquals("/files/100% done.pdf", ClientCacheFilter.canonicalize("/files/100% done.pdf"));
+        assertEquals("/files/50% off.pdf", ClientCacheFilter.canonicalize("/files/50%25 off.pdf"));
+        // A plus sign is a plus sign in a path, whatever form encoding makes of it.
+        assertSame("/files/a+b.pdf", ClientCacheFilter.canonicalize("/files/a+b.pdf"));
+    }
+
+    @Test
+    public void canonicalizeDoesNotClimbAboveTheRoot() {
+        assertEquals("/etc/passwd", ClientCacheFilter.canonicalize("/../../etc/passwd"));
+        assertEquals("/", ClientCacheFilter.canonicalize("/"));
     }
 
     /** An {@link HttpServletRequest} that answers only the three methods this reads. */
